requirements.txt
ynab>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
